Remove ptr-int-ptr casts
This replaces ptr-int-ptr roundtrips as well as ptr-int casts which were
used to produce usize values to do arithmetic on instead of doing
pointer arithmetic operations on pointers to u8. These ptr-int-ptr
roundtrips are a thorny case for aliasing models to deal with, and thus
play havoc with unsafe checkers such as Miri, and complicate alias
analysis and aliasing models. Removing them does not necessarily remove
UB, but it significantly simplifies the semantics of this code and makes
it easier to rule out definite forms of UB.